Pluggables; ScriptableO and MonoB
Objects that Baxter can plug cables into are called pluggables.
To help make them compatible with the recipe system, they have a two-fold structure.

Their data is held in a scriptable object called a 'PluggableSO'. This includes their sprite and collider dimensions.
The recipe system can use arrays of PluggableSO's to define its recipes.
There is boilerplate code in there for when the object is plugged/unplugged.
These have some child classes for speakers and midaffectors so they can have special parameters and functionality.

The other object is a gameobject with a monobehaviour script called 'PluggableMB'.
Each of these needs a PluggableSO associated with it to define it.
They have a sprite renderer that finds its sprite from its PluggableSO, and much the same regarding a box collider.
There is a prefab blank one of these, that you can bring into a scene before initialising it by setting a PluggableSO for it.