Instrument SO and MB
Instruments use a scriptable object and monobehaviour system much like pluggables.
However these are controlled more by the monobehaviour than the SO, which is the opposite of in the case of pluggables.
Set up instruments by creating a scriptable object and filling in its data, then applying this to intrument monobehaviour object in the scene (you can use the prefab and unpack it completely).
If an instrument is a drum, it should play from the start of each song without requiring a speaker because they r loud