PlugCables now have a static colour & PlugSOs done
PROBLEM - You cannot see the correct sprite for PlugCables in the editor before playmode anymore
Improvements:
You can set the colour of a PlugCable item in inspector, it is an enum value.
You need to set a value on the instrumentMB to choose the colour of the cable coming out of it.
PluggableSOs now take four sprites, for the four colours and choose their colour on game start.